More house keeping and trimming of unused functions
