+++ porespy/visualization/__init__.py
@@ -1,2 +1,1 @@
-from .__views__ import sem, xray, local_thickness
-from .__plots__ import drainage_curve
+from .__views__ import sem, xray
